Add authorization to all of the api routes

+++ pages/api/delete/volum.ts
@@ -6,16 +6,27 @@ const prisma = new PrismaClient();
 
 export default async function handler(
   req: NextApiRequest,
-  res: NextApiResponse
+  res: NextApiResponse<Volum | { message: string }>
 ) {
-  const idVolumSaFieSters: string = req.body.id;
+  try {
+    if (
+      req.headers["security-phrase"] === process.env.NEXT_PUBLIC_SECURITY_PHRASE
+    ) {
+      const idVolumSaFieSters: string = req.body.id;
 
-  await prisma.volume.delete({
-    where: {
-      id: idVolumSaFieSters,
-    },
-  });
+      const volumSters = await prisma.volume.delete({
+        where: {
+          id: idVolumSaFieSters,
+        },
+      });
 
-  res.status(200).json("Volum sters cu succes!");
-  prisma.$disconnect();
+      res.status(200).json(volumSters);
+    } else {
+      res.status(401).json({ message: "Unauthorized" });
+    }
+  } catch (error: any) {
+    res.status(500).json(error);
+  } finally {
+    prisma.$disconnect();
+  }
 }
